[#0] restructure backend directories to better match clean architecture
