+++ backend/api/src/main/kotlin/software/shonk/adapters/incoming/HelloWorldController.kt
@@ -3,7 +3,7 @@ package software.shonk.adapters.incoming
 import io.ktor.server.response.respondText
 import io.ktor.server.routing.Route
 import io.ktor.server.routing.get
-import software.shonk.application.port.HelloWorldUseCase
+import software.shonk.application.port.incoming.HelloWorldUseCase
 
 // Configure all endpoints regarding Hello World in here (example)
 fun Route.configureHelloWorldController(helloWorldUseCase: HelloWorldUseCase) {
