+++ backend/api/src/main/kotlin/software/shonk/application/port/incoming/HelloWorldUseCase.kt
@@ -1,4 +1,4 @@
-package software.shonk.application.port
+package software.shonk.application.port.incoming
 
 interface HelloWorldUseCase {
     fun sayHello(): String
